+++ src/main/java/com/imsweb/algorithms/ruralurban/RuralUrbanDataProvider.java
@@ -6,9 +6,9 @@ package com.imsweb.algorithms.ruralurban;
 /**
  * The purpose of the <code>RuralUrbanContinuumDataProvider</code> is to:
  * <ul>
- * <li>get the rural urban continuum code for the provided year category, state of dx, and county of dx</li>
- * <li>get the rural urban commuting area code for the provided year category, state of dx, county of dx, and census tract</li>
- * <li>get the rural urban census code for the provided year category, state of dx, county of dx, and census tract</li>
+ * <li>get the rural urban continuum code for the provided year category, state of dx, and county of dx for analysis</li>
+ * <li>get the rural urban commuting area code for the provided year category, state of dx, county of dx for analysis, and census tract</li>
+ * <li>get the rural urban census code for the provided year category, state of dx, county of dx for analysis, and census tract</li>
  * </ul>
  * either from the database or csv lookup based on the implementation.
  * <p/>
@@ -25,45 +25,45 @@ public interface RuralUrbanDataProvider {
     String TRACT_CATEGORY_2 = "2010";
 
     /**
-     * Returns urban rural indicator code (URIC) for provided year category, state of dx, county of dx, and census tract.
+     * Returns urban rural indicator code (URIC) for provided year category, state of dx, county of dx for analysis, and census tract.
      * <p/>
      * Created Aug 11, 2014 by HoweW
      * @param state state at DX
-     * @param county county at DX
+     * @param county county at DX for analysis
      * @param tractCategory census tract
      * @return the corresponding rural urban census code
      */
     String getUrbanRuralIndicatorCode(String tractCategory, String state, String county, String censusTract);
 
     /**
-     * Returns urban rural indicator code (URIC) percentage for provided year category, state of dx, county of dx, and census tract.
+     * Returns urban rural indicator code (URIC) percentage for provided year category, state of dx, county of dx for analysis, and census tract.
      * <p/>
      * Created Feb 28, 2015 by depryf
      * @param state state at DX
-     * @param county county at DX
+     * @param county county at DX for analysis
      * @param tractCategory census tract
      * @return the corresponding rural urban census percentage
      */
     Float getRuralUrbanCensusPercentage(String tractCategory, String state, String county, String censusTract);
 
     /**
-     * Returns rural urban commuting area (RUCA) code for provided year category, state of dx, county of dx, and census tract.
+     * Returns rural urban commuting area (RUCA) code for provided year category, state of dx, county of dx for analysis, and census tract.
      * <p/>
      * Created Aug 11, 2014 by HoweW
      * @param state state at DX
-     * @param county county at DX
+     * @param county county at DX for analysis
      * @param tractCategory census tract
      * @return the corresponding rural urban commuting area code
      */
     String getRuralUrbanCommutingArea(String tractCategory, String state, String county, String censusTract);
 
     /**
-     * Returns rural urban continuum code for provided year category, state of dx, and county of dx.
+     * Returns rural urban continuum code for provided year category, state of dx, and county of dx for analysis.
      * <p/>
      * Created Aug 11, 2014 by HoweW
      * @param bealeCategory year of beale
      * @param state state at DX
-     * @param county county at DX
+     * @param county county at DX for analysis
      * @return the corresponding rural urban continuum code
      */
     String getRuralUrbanContinuum(String bealeCategory, String state, String county);